requirements.txt
requests>=2.31.0
python-dateutil>=2.8.0
pytest>=7.0.0
